requirements.txt
# Discord Bot Dependencies
discord.py==2.3.2
wavelink==3.4.1

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Web Framework for API
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.8.2
yarl==1.10.0
pydantic>=2.0.0

# System utilities
psutil==5.9.6
colorama==0.4.6

# Logging and monitoring
loguru==0.7.2

# Node communication
redis==5.0.1
aioredis==2.0.1

# Security
cryptography>=41.0.0

# YouTube support
yt-dlp>=2025.7.21